// backend/src/notifications/enums/notification-target-entity.enum.ts
export enum NotificationTargetEntity {
  RM_REQUEST = 'RM_REQUEST',
  MATERIAL_ISSUE = 'MATERIAL_ISSUE',
  ADDITIONAL_MATERIAL_REQUEST = 'ADDITIONAL_MATERIAL_REQUEST',
  ADDITIONAL_REQUEST = 'ADDITIONAL_REQUEST',
  SC = 'SC',
  DELIVERY_CHALLAN = 'DELIVERY_CHALLAN',
  MATERIAL_RECEIPT = 'MATERIAL_RECEIPT',
  MATERIAL_RETURN = 'MATERIAL_RETURN',
}